Integrate network metrics collection
