+++ peerNode/client.py
@@ -1,11 +1,9 @@
 from jsonrpclib import Server
 import sys
 from datetime import datetime
-#from metrics import get_jitter, get_latency
-#from discovery import get_cluster_info, save_cluster_info, generate_cluster_info
 from util import retrieve_clusters_info, check_availability, check_cluster_resources
+from network import check_network_resources
 import requests
-#from metrics import get_throughtput
 import os
 import socket
 import time
@@ -14,6 +12,7 @@ hostname = socket.gethostname()
 CARBON_SERVER = socket.gethostbyname(hostname)
 CARBON_PORT = int(os.getenv('CARBON_PORT'))
 DELAY = int(os.getenv("DELAY"))
+IPERF = int(os.getenv("IPERF"))
 
 
 def push_to_graphite(metrics):
@@ -38,6 +37,7 @@ def main():
             if availability == 1:
 
                 resources = check_cluster_resources(cluster["ip_address"], cluster["port"])
+                network = check_network_resources(cluster["ip_address"], IPERF)
 
                 if resources:
                     resource_message = [
@@ -48,55 +48,19 @@ def main():
                     graphite_r_message = '\n'.join(resource_message) + '\n'    
                     push_to_graphite(graphite_r_message)
 
-            #lines = [
-            #'%s %s %s %d %d' % (cluster["cluster_id"], "Availability", "A", availability, int(time.time()))
-            #'system.%s.loadavg_5min %s %d' % (node, loadavgs[1], timestamp),
-            #'system.%s.loadavg_15min %s %d' % (node, loadavgs[2], timestamp)
-            #]
+                if network:
+                    network_message = [
+                        '%s %s %s %d %d' % (cluster["cluster_id"], "Network", "T", network["throughput"], int(time.time())),
+                        '%s %s %s %d %d' % (cluster["cluster_id"], "Network", "L", network["latency"], int(time.time())),
+                        '%s %s %s %d %d' % (cluster["cluster_id"], "Network", "J", network["jitter"], int(time.time()))
+                        ]
+                    graphite_n_message = '\n'.join(network_message) + '\n'    
+                    push_to_graphite(graphite_n_message)
 
-            #if availability == 1:
-            #    resources = get_cluster_resources(cluster["ip_address"], cluster["port"])
-            #    network = get_network_resources(cluster["ip_address"], cluster["port"])
-        
-            #else:
-
-            #    return None
             availability_metric = '%s %s %s %d %d\n' % (cluster["cluster_id"], "Availability", "A", availability, int(time.time()))
             push_to_graphite(availability_metric)
         
         time.sleep(DELAY)
 
-
-#    port = 5001
-#    server = Server(f'http://localhost:{port}')
-
-#    try:
-#        print("Get server metrics")
-#        print(server.get_server_resources())
-
-#    except:
-#        print("Error: ", sys.exc_info())
-
-#def store_metrics(server, host, port, cluster_id):
-#   latency = get_latency(host,port)
-#    jitter = get_jitter(host,port)
-#    date = datetime.date
-#    print("Get throughput.....")
-#    throughput = get_throughtput(host)
-#    print(throughput)
-#    metrics = (cluster_id, throughput,latency,jitter, str(date))
-#    print(metrics)
-#    print("Inserting network data: latency, throughput & jitter..")
-#    resp = server.insert_network(metrics)
-#    print(resp)
-
-    # insert availability metrics
-#    print("Handle availability....")
-#    availability_score = check_availability(host,port)
-#    print(availability_score)
-#    metrics = (cluster_id, availability_score, str(date))
-#    print(metrics)
-#    print(server.insert_availability(metrics))
-
-#if __name__ == '__main__':
-#    main()
+if __name__ == '__main__':
+    main()
